requirements.txt
discord.py
python-dotenv
pymongo==3.12.0